requirements.txt
pandas==1.0.1
numpy==1.18.1
Pillow==7.0.0
flickrapi==2.4.0
urllib3==1.25.8
LightGBM==2.2.1
torch==0.4.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
sklearn==0.0
category-encoders
tqdm==4.23.0
gensim==3.8.1
nltk==3.4.5
opencv-contrib-python==4.2.0.32
tensorflow
matplotlib==3.1.3
catboost